Move accessors to separate class called MemoryAccess
